+++ Sources/Platform/Linux/LinuxSystemTray.swift
@@ -1,0 +1,51 @@
+// Sources/Platform/Linux/LinuxSystemTray.swift
+// Linux implementation of system tray using StatusNotifierItem protocol
+
+#if os(Linux)
+
+import Foundation
+
+/// Linux system tray implementation using D-Bus StatusNotifierItem
+///
+/// This is a placeholder implementation that will eventually communicate
+/// with the system tray via D-Bus using the StatusNotifierItem protocol.
+///
+/// References:
+/// - https://www.freedesktop.org/wiki/Specifications/StatusNotifierItem/
+public final class LinuxSystemTray: SystemTray {
+    private var currentIcon: PlatformImage?
+    private var currentMenu: PlatformMenu?
+    private var visible: Bool = false
+
+    public init() {
+        // TODO: Initialize D-Bus connection
+        // TODO: Register StatusNotifierItem on session bus
+    }
+
+    public func setIcon(_ image: PlatformImage) {
+        currentIcon = image
+        // TODO: Update icon via D-Bus IconPixmap or IconName property
+    }
+
+    public func setMenu(_ menu: PlatformMenu) {
+        currentMenu = menu
+        // TODO: Export menu via DBusMenu protocol
+        // TODO: Set Menu property on StatusNotifierItem
+    }
+
+    public func show() {
+        visible = true
+        // TODO: Make StatusNotifierItem visible via D-Bus
+    }
+
+    public func hide() {
+        visible = false
+        // TODO: Hide StatusNotifierItem via D-Bus
+    }
+
+    public var isVisible: Bool {
+        visible
+    }
+}
+
+#endif